requirements.txt
selenium
python-telegram-bot
pymongo
